Low: Build: Determin the correct format for printing uint64_t so that it need not be cast
Mercurial revision: a6fb0aae90c19803abe5dc775c7715be87e0c77b

+++ include/crm/ais_common.h
@@ -119,9 +119,9 @@ enum crm_proc_flag {
 
 typedef struct crm_peer_node_s 
 {
-	unsigned int id;
-	unsigned long long born;
-	unsigned long long last_seen;
+	uint32_t id;
+	uint64_t born;
+	uint64_t last_seen;
 
 	int32_t votes;
 	uint32_t processes;
